Music / Sound Effect Control
Added in functionality to the music and sound effect checkboxes found
on the settings menu. Checked = on, unchecked = off.

Also added in the ability for the game to automatically have the
checkboxes selected / unselected when the scene is pushed depending on
whether or not the music is on or off.