Possible Particle Fix?
So I changed it that when it enters turbo mode the particle blend
becomes normal and when you exit turbo mode, it switches back to the
additive. I think this gives the best of both worlds. It still looks a
little clunky but i think it’s a good solution. Lemme know what you
think. I also fixed the error with the innerParticleEmissionRate is not
defined.